+++ movable_objects/boom.cpp
