+++ movable_objects/bot.cpp
