+++ movable_objects/cleverbot.cpp
