+++ movable_objects/cleverbot.h
@@ -1,12 +1,12 @@
-#ifndef MOVABLES_CLEVERBOT_H_
-#define MOVABLES_CLEVERBOT_H_
+#ifndef MOVABLE_OBJECTS_CLEVERBOT_H_
+#define MOVABLE_OBJECTS_CLEVERBOT_H_
 
 #include <QQueue>
 #include <QVector>
 #include <memory>
 #include <vector>
 #include "improvedbot.h"
-#include "../map_objects/portal.h"
+#include "../static_objects/portal.h"
 
 class CleverBot : public ImprovedBot {
  public:
@@ -38,4 +38,4 @@ class CleverBot : public ImprovedBot {
   int width_;
 };
 
-#endif  // MOVABLES_CLEVERBOT_H_
+#endif  // MOVABLE_OBJECTS_CLEVERBOT_H_
