+++ movable_objects/improvedbot.cpp
