+++ movable_objects/rocket.cpp
