+++ movable_objects/tank.cpp
