+++ static_objects/objectonmap.cpp
