+++ static_objects/objectonmap.h
@@ -1,5 +1,5 @@
-#ifndef MAP_OBJECTS_OBJECTONMAP_H_
-#define MAP_OBJECTS_OBJECTONMAP_H_
+#ifndef STATIC_OBJECTS_OBJECTONMAP_H_
+#define STATIC_OBJECTS_OBJECTONMAP_H_
 
 #include <QDebug>
 #include <QImage>
@@ -51,4 +51,4 @@ class Charge : public ObjectOnMap {
   Charge(const std::shared_ptr<Map>& map, size_t x, size_t y);
 };
 
-#endif  // MAP_OBJECTS_OBJECTONMAP_H_
+#endif  // STATIC_OBJECTS_OBJECTONMAP_H_
