+++ static_objects/portal.cpp
